requirements.txt
kubernetes==29.0.0
prometheus-api-client==0.5.5
numpy==1.26.4
pyyaml==6.0.1
requests==2.31.0
urllib3==2.2.0
